requirements.txt
# Core backend
Flask==3.1.0
flask-cors==5.0.1
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1

# ML & Image Processing
numpy==1.24.4
opencv-python==4.8.0.76
pandas==2.2.2
Pillow==9.5.0
scikit-learn==1.3.2
xgboost==2.0.3

# Deep learning (custom models)
tensorflow==2.13.0
keras==2.13.1

# Azure Storage (if loading models from Blob)
azure-storage-blob==12.19.1
azure-identity

# For safely loading PyTorch YOLOv5 models if needed
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1

# Utilities
gunicorn==21.2.0
